refactor: extract buttons to reusable components & separate types, reducer
