+++ styles/global.css
@@ -26,6 +26,6 @@ img {
   display: block;
 } */
 
-/* @tailwind base;
+@tailwind base;
 @tailwind components;
-@tailwind utilities; */
+@tailwind utilities;
